Add scram.py for python3, rename the python2 one to scram2.py
In both versions, the recursive hi function is changed to an iterative function. Some changes were made in both files to ensure they use as similar code as they can. There must be an issue with the encodings in the python3 version, as some functions that are not backwards-compatible expect not string type arguments but bytes/bytearray type arguments. I am yet to figure out how to fix this.